+++ test/PhpSessionPersistenceFactoryTest.php
@@ -27,6 +27,7 @@ class PhpSessionPersistenceFactoryTest extends TestCase
         $persistence = $factory($container->reveal());
         $this->assertInstanceOf(PhpSessionPersistence::class, $persistence);
         $this->assertFalse($persistence->isNonLocking());
+        $this->assertFalse($persistence->isDeleteCookieOnEmptySession());
 
         // test php-session-persistence with non-locking config set to false and true
         foreach ([false, true] as $nonLocking) {
@@ -42,6 +43,22 @@ class PhpSessionPersistenceFactoryTest extends TestCase
             ]);
             $persistence = $factory($container->reveal());
             $this->assertSame($nonLocking, $persistence->isNonLocking());
+        }
+
+        // test php-session-persistence with delete_cookie_on_empty_session config set to false and true
+        foreach ([false, true] as $deleteCookieOnEmptySession) {
+            $container->has('config')->willReturn(true);
+            $container->get('config')->willReturn([
+                'session' => [
+                    'persistence' => [
+                        'ext' => [
+                            'delete_cookie_on_empty_session' => $deleteCookieOnEmptySession,
+                        ],
+                    ],
+                ],
+            ]);
+            $persistence = $factory($container->reveal());
+            $this->assertSame($deleteCookieOnEmptySession, $persistence->isDeleteCookieOnEmptySession());
         }
     }
 }